--- a/src/main/java/org/shanerx/tradeshop/listeners/ShopProtectionListener.java
+++ b/src/main/java/org/shanerx/tradeshop/listeners/ShopProtectionListener.java
@@ -138,6 +138,8 @@ public class ShopProtectionListener extends Utils implements Listener {
         event.setCancelled(hopperEvent.isForbidden());
         plugin.getListManager().addSkippableHopper(event.getInitiator().getLocation(), hopperEvent.isForbidden());
         debugger.log("ShopProtectionListener: (TSAF) HopperEvent isForbidden: " + hopperEvent.isForbidden(), DebugLevels.PROTECTION);
+        shop.updateStatus();
+        shop.saveShop();
     }
 
     @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)