--- a/src/main/java/org/shanerx/tradeshop/listeners/ShopProtectionListener.java
+++ b/src/main/java/org/shanerx/tradeshop/listeners/ShopProtectionListener.java
@@ -139,6 +139,7 @@ public class ShopProtectionListener extends Utils implements Listener {
         plugin.getListManager().addSkippableHopper(event.getInitiator().getLocation(), hopperEvent.isForbidden());
         debugger.log("ShopProtectionListener: (TSAF) HopperEvent isForbidden: " + hopperEvent.isForbidden(), DebugLevels.PROTECTION);
         shop.updateStatus();
+        shop.updateSign();
         shop.saveShop();
     }