… pom.xml
… src/main/java/org/shanerx/tradeshop/objects/ListManager.java
… src/main/java/org/shanerx/tradeshop/objects/Shop.java
… src/main/java/org/shanerx/tradeshop/utils/Utils.java
… src/main/java/org/shanerx/tradeshop/utils/config/ConfigManager.java
… src/main/java/org/shanerx/tradeshop/utils/config/Message.java
… src/main/java/org/shanerx/tradeshop/utils/config/Setting.java
… src/main/resources/Lang/en-us.yml