@KillerOfPie 2022-03-29
> Were the contents of the list returned by `getItems()` tested?

I did basic testing to make sure it worked and will do more once I update the missing items messages.