@KillerOfPie 2022-03-29
What do you think of adding a Guava Cache for items in DataStorage, they would save normally but we would lower the amount of disk reads we do. Also since they aren't read as ofen or quick as the Hoppers the Caches can max at like 100 and expireAfterAccess at 1-5min?