@KillerOfPie 2022-03-31
got bored so I implemented anyways... Going to push to a separate branch so you can review and decide if we should add in or not. See #125 and leave further discussion/review there.

Below is my basic testing for it

The Red lines are the Approximate start/end of the tests

## Trade Testing
Test involves entering server, trading 5 times then returning the items for each of 10 self owned shops, then screenshot.

### Current 2.5.1 Branch:
![image](https://user-images.githubusercontent.com/7243280/160979317-4bde0792-805f-4cf8-b7ba-d76a319d30e3.png)

### With new Caching Code:
![image](https://user-images.githubusercontent.com/7243280/160979784-e5e7d469-0a49-45bf-a191-6b04158381ca.png)

## Hopper Testing
Test involves entering world and waiting for mspT(ms per Tick) to stabilize then teleporting to my hopper testing area with 80k hoppers and again waiting for mspT to stabilize. This will test if checking for shops from the storage is faster with the linkage caching.

### Current 2.5.1 Branch:
Pre-Teleport Avg mspT: ~3.3ms
Post-Teleport Avg mspT: ~78ms
Post-Teleport Peak mspT: ~150ms
Return Avg mspT: ~2.8ms

Post Test Screenshot:
Forgot to hover for peak and looks like it just GC'd, peaks looks to be around 4000mb
![image](https://user-images.githubusercontent.com/7243280/160980976-de2441ee-1e4e-46f5-bd27-0d78deaf6eb6.png)

Post Return Screenshot:
![image](https://user-images.githubusercontent.com/7243280/160981259-58c3a3cb-d361-4e83-a60e-99ae0d49ff95.png)


### With new Caching Code:
Pre-Teleport Avg mspT: ~3.1ms
Post-Teleport Avg mspT: ~ms
Post-Teleport Peak mspT: ~500ms+
Return Avg mspT: ~ms

Ended before full Peak was reached.
This probably should've been obvious since caching only helps re-accessing the items and the hoppers would have to each read from disk to check since they wouldn't be in the cache. Only option I see around this is keeping all the linkage data in memory(would have to check each world folder for linkage data on startup and load them all into memory which i'm not a fan of since it wouldn't get cleaned out of memory.